8319541: G1: Inline G1RemoveSelfForwardsTask into RestoreRetainedRegionsTask
